Create global composables file and tests
