+++ test/unit/HomeEpisode.spec.ts
@@ -1,76 +1,40 @@
-import { useAudio, useSendScore } from "@/components/HomeEpisode";
+import { useSendScore } from "@/composables/HomeEpisode";
 
 describe("HomeEpisode unit tests", () => {
-  describe("useAudio composable", () => {
-    test("With audio element", () => {
-      let playCount = 0;
-      let pauseCount = 0;
+  describe("useSendScore composable", () => {
+    test("With ranking id", () => {
+      const rankingId = "123sdsf";
+      let eventName = "";
+      let eventValues: Array<string> = [];
 
-      const { playing, play } = useAudio({
-        play: () => {
-          playCount++;
-        },
-        pause: () => {
-          pauseCount++;
-        },
-      } as any);
+      const emit = (name: string, vals: Array<string>) => {
+        eventName = name;
+        eventValues = vals;
+      };
 
-      expect(playing.value).toBe(false);
-      expect(playCount).toBe(0);
-      expect(pauseCount).toBe(0);
-      play();
-      expect(playing.value).toBe(true);
-      expect(playCount).toBe(1);
-      expect(pauseCount).toBe(0);
-      play();
-      expect(playing.value).toBe(false);
-      expect(playCount).toBe(1);
-      expect(pauseCount).toBe(1);
+      const { sendScore } = useSendScore(emit);
+
+      sendScore(rankingId);
+
+      expect(eventName).toBe("show-score");
+      expect(eventValues[0]).toBe(rankingId);
     });
 
-    test("Without audio element", () => {
-      const { playing, play } = useAudio(null as any);
+    test("Without ranking id", () => {
+      let eventName = "";
+      let eventValues: Array<string> = [];
 
-      expect(playing.value).toBe(false);
-      play();
-      expect(playing.value).toBe(false);
-    });
+      const emit = (name: string, vals: Array<string>) => {
+        eventName = name;
+        eventValues = vals;
+      };
 
-    describe("useSendScore composable", () => {
-      test("With ranking id", () => {
-        const rankingId = "123sdsf";
-        let eventName = "";
-        let eventValues: Array<string> = [];
+      const { sendScore } = useSendScore(emit);
 
-        const emit = (name: string, vals: Array<string>) => {
-          eventName = name;
-          eventValues = vals;
-        };
+      sendScore(undefined);
 
-        const { sendScore } = useSendScore(emit);
-
-        sendScore(rankingId);
-
-        expect(eventName).toBe("show-score");
-        expect(eventValues[0]).toBe(rankingId);
-      });
-
-      test("Without ranking id", () => {
-        let eventName = "";
-        let eventValues: Array<string> = [];
-
-        const emit = (name: string, vals: Array<string>) => {
-          eventName = name;
-          eventValues = vals;
-        };
-
-        const { sendScore } = useSendScore(emit);
-
-        sendScore(undefined);
-
-        expect(eventName).toBe("");
-        expect(eventValues[0]).toBe(undefined);
-      });
+      expect(eventName).toBe("");
+      expect(eventValues[0]).toBe(undefined);
     });
   });
 });
